feat(client): 👜 add react setup
