+++ package.json
@@ -15,9 +15,13 @@
     "commander": "^14.0.0",
     "express": "^4.18.2",
     "isomorphic-git": "^1.21.0",
-    "matter-js": "^0.20.0"
+    "matter-js": "^0.20.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0"
   },
   "devDependencies": {
+    "@types/react": "^19.1.8",
+    "@types/react-dom": "^19.1.6",
     "@types/express": "^5.0.3",
     "@types/jest": "^29.5.14",
     "@types/matter-js": "^0.19.8",
