+++ public/index.html
@@ -172,10 +172,13 @@
     <div id="timestamp"></div>
     <div id="sim"></div>
     <div id="commit-log"></div>
+    <div id="root"></div>
     <script type="importmap">
       {
         "imports": {
-          "matter-js": "https://cdn.jsdelivr.net/npm/matter-js@0.20.0/+esm"
+          "matter-js": "https://cdn.jsdelivr.net/npm/matter-js@0.20.0/+esm",
+          "react": "https://cdn.jsdelivr.net/npm/react@18.2.0/+esm",
+          "react-dom/client": "https://cdn.jsdelivr.net/npm/react-dom@18.2.0/client/+esm"
         }
       }
     </script>
