+++ tsconfig.client.json
@@ -6,7 +6,8 @@
     "module": "ESNext",
     "target": "ES2020",
     "moduleResolution": "Bundler",
-    "verbatimModuleSyntax": true
+    "verbatimModuleSyntax": true,
+    "jsx": "react-jsx"
   },
-  "include": ["src/client/index.ts"]
+  "include": ["src/client/index.tsx"]
 }
